requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.44.0
huggingface_hub>=0.22.2
accelerate>=0.33.0
Pillow>=10.3.0
numpy>=1.26.4
scikit-learn>=1.4.2
umap-learn>=0.5.6
plotly>=5.23.0
dash>=2.17.0
dash-bootstrap-components>=1.6.0
PyYAML>=6.0.1
opencv-python-headless>=4.10.0.84
tqdm>=4.66.4

